fix(wake): define snapshot error in cross-platform file
