+++ internal/cli/wake_lifecycle_guard_unix.go
@@ -21,28 +21,6 @@ type wakeAgentDir struct {
 	closed bool
 }
 
-// wakeSnapshotReadChangedError marks an observation that could not bind one
-// stable file identity. It is retryable only when the caller can restart the
-// complete read-only decision without reusing evidence from the failed read.
-type wakeSnapshotReadChangedError struct {
-	err error
-}
-
-func (err *wakeSnapshotReadChangedError) Error() string {
-	return err.err.Error()
-}
-
-func (err *wakeSnapshotReadChangedError) Unwrap() error {
-	return err.err
-}
-
-func newWakeSnapshotReadChangedError(err error) error {
-	if err == nil {
-		return nil
-	}
-	return &wakeSnapshotReadChangedError{err: err}
-}
-
 func wakeLifecycleGuardPath(root, me string) string {
 	return filepath.Join(fsq.AgentBase(root, me), wakeLifecycleGuardFileName)
 }
